+++ src/components/UI/LanguageSelect/LanguageSelect.tsx
@@ -1,17 +1,26 @@
 import styled from 'styled-components';
 import Language from "./Language/Language";
+import {Locale} from "../../../localization/types";
 
 type LanguageSelectProps = {
-    languages: Array<{name: string, englishName: string, locale: string}>,
-    selectLocale: (locale: string) => void
+    languages: Array<{name: string, englishName: string, locale: Locale}>,
+    selectLocale: (locale: Locale) => void,
+    selectedLocale: Locale
 }
 
 const StyledContainer = styled.div`
-  min-width: 15rem;
-  max-height: 40rem;
+  min-width: 100%;
+  min-height: 100%;
+  height: 100%;
   border-radius: 5px;
-  background-color: white;
   overflow-y: auto;
+    () => {
+      -webkit-touch-callout: none; /* iOS Safari */
+      -webkit-user-select: none; /* Safari */
+      -moz-user-select: none; /* Old versions of Firefox */
+      -ms-user-select: none; /* Internet Explorer/Edge */
+      user-select: none; /* Non-prefixed version, currently supported by Chrome, Edge, Opera and Firefox */
+    }
 `
 
 const LanguageSelect = (props: LanguageSelectProps) => {
@@ -23,6 +32,7 @@ const LanguageSelect = (props: LanguageSelectProps) => {
                     englishName={language.englishName}
                     locale={language.locale}
                     selectLocale={props.selectLocale}
+                    selectedLocale={props.selectedLocale}
                     key={idx}/>
             )}
         </StyledContainer>